make output in different folder

+++ Analysis/Mutation/Mutation_analysis.sh
@@ -6,7 +6,9 @@ base_dir=$3
 
 species_analysis="$species""_""$analysis"
 
-rm -r "./$species_analysis/"
+if [ -d "$species_analysis" ]; then
+	rm -r "./$species_analysis/"
+fi
 
 mkdir "./$species_analysis/"
